#65 move methods to category files
