+++ internal/repository/category.go
@@ -1,6 +1,8 @@
 package repository
 
 import (
+	"database/sql"
+	"errors"
 	"fmt"
 
 	"github.com/MAVIKE/yad-backend/internal/domain"
@@ -73,4 +75,73 @@ func (r *CategoryPg) GetAllItems(categoryId int) ([]*domain.MenuItem, error) {
 	err := r.db.Select(&items, query, categoryId)
 
 	return items, err
+}
+
+func (r *CategoryPg) DeleteCategory(restaurantId int, categoryId int) error {
+	tx, err := r.db.Begin()
+	if err != nil {
+		return err
+	}
+
+	id := 0
+	query := fmt.Sprintf(`SELECT id FROM %s WHERE restaurant_id = $1 AND id = $2`, categoriesTable)
+	row := r.db.QueryRow(query, restaurantId, categoryId)
+	err = row.Scan(&id)
+	if err != nil {
+		if err == sql.ErrNoRows {
+			_ = tx.Rollback()
+			return errors.New("category does not belong to this restaurant")
+		} else {
+			_ = tx.Rollback()
+			return err
+		}
+	}
+
+	query = fmt.Sprintf(`DELETE FROM %s WHERE category_id = $1`, categoryItemsTable)
+	_, err = r.db.Exec(query, categoryId)
+	if err != nil {
+		_ = tx.Rollback()
+		return err
+	}
+
+	query = fmt.Sprintf(`DELETE FROM %s WHERE id = $1`, categoriesTable)
+	_, err = r.db.Exec(query, categoryId)
+	if err != nil {
+		_ = tx.Rollback()
+		return err
+	}
+
+	return tx.Commit()
+}
+
+func (r *CategoryPg) UpdateCategory(restaurantId int, categoryId int, input *domain.Category) error {
+	tx, err := r.db.Begin()
+	if err != nil {
+		return err
+	}
+
+	id := 0
+	query := fmt.Sprintf(`SELECT id FROM %s WHERE restaurant_id = $1 AND id = $2`, categoriesTable)
+	row := r.db.QueryRow(query, restaurantId, categoryId)
+	err = row.Scan(&id)
+	if err != nil {
+		if err == sql.ErrNoRows {
+			_ = tx.Rollback()
+			return errors.New("category does not belong to this restaurant")
+		}
+
+		_ = tx.Rollback()
+		return err
+	}
+
+	if input.Title != "" {
+		query = fmt.Sprintf(`UPDATE %s SET title = $1 WHERE id = $2`, categoriesTable)
+		_, err = r.db.Exec(query, input.Title, categoryId)
+		if err != nil {
+			_ = tx.Rollback()
+			return err
+		}
+	}
+
+	return tx.Commit()
 }
